Surgery Add Event Changes
Corrected an issue in the stored procedure surgAddEvent, which did not
properly account for unscheduled rooms.
the SurgeryRoomAdd page now allows for multiple selection of both
providers and items, which are correctly stored on the db.
TODO: Recreate Surgery Events table & properly display both providers
and items in the Surgery Viewer.