Misc. Sessions & Surgery Add Utility Changes
Changed start time to 6:30 am. Added additional session checks at the
start of the Surgery Viewer. Replaced redirect queries with sessions.
Added session checks to the Surgery Add page. Added postback event to
maintain previous item focus in the event of a postback. Reinstated the
LoadPatients method to allow a "lookup" ListBox of patients; this
ListBox does nothing to the add query, it simply reflects the patient
name of the value entered in the textfield to its left. This ListBox
should also not be focused when using the tab key.