Main Surgery Table Alterations, + Misc.
Redid the main schedule table, so currently there are very few stored
surgery appointments.
Created a table in the BSU_PROJECT db for the weight fields. Altered the
query to pull this field as demonstrated in the email.
Redid a lot of the basicInfoSurgery to make it less confusing to
navigate, replacing integer references to column name references.
Altered the drag and drop query to (hopefully) properly work; the
additional constraints were causing issues.
Added a table for the OR Rooms. Changing the location also resets the
selected item of the OR rooms for both the main Surgery Viewer and the
Add Surgery page.
Added a stored procedure to delete surgery events.
Added stored proedures to fill in Add Surgery page with the details of
an event to modify. Currently passing through an unused Session
variable, but this is not passed, as there is currently no means to
actually select an event to edit from the Surgery Viewer.
Began validation on the Add Surgery page, validating the duration to be
within the available time frame of the or rooms.